requirements.txt
requests==2.24.0
firebase==3.0.1
google-cloud-storage==1.31.2
urllib3==1.25.10
https://download.pytorch.org/whl/cpu/torch-1.3.1%2Bcpu-cp37-cp37m-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.4.2%2Bcpu-cp37-cp37m-linux_x86_64.whl
scikit-image==0.14.2
scipy==1.1.0
argparse
numpy==1.15.0
pytesseract
tesseract
pyproj
psycopg2
flask
flask_restful
flask_cors
opencv-python
python_jwt
gcloud
sseclient
pycrypto
requests-toolbelt